apps/workers/transcription/requirements.txt
faster-whisper==1.0.1
pyannote.audio==3.1.1
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
supabase==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
